custom-nodes/node-red-contrib-azurellmnode/requirements.txt
azure-ai-formrecognizer>=3.2.0
azure-search-documents>=11.4.0
openai>=1.12.0
python-dotenv>=0.19.0